requirements.txt
torch~=2.0.1
torchvision~=0.15.2
torchaudio~=2.0.2
torch-summary~=1.4.5
numpy~=1.24.3
opencv-python~=4.7.0.72
tqdm~=4.65.0
scipy~=1.10.1
scikit-learn~=1.2.2
Pillow~=9.5.0
pandas~=2.0.1
matplotlib~=3.7.1